% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_package.R
\docType{data}
\name{ex_dta_rct}
\alias{ex_dta_rct}
\title{Example dataset}
\format{
A data frame with the following variables:

\describe{
  \item{Group}{current, rwd}
  \item{Arm}{control, treatment}
  \item{Y_Con}{Continuous outcome}
  \item{V1}{Baseline covariate}
  \item{V2}{Baseline covariate}
  \item{V3}{Baseline covariate}
  \item{V4}{Baseline covariate}
  \item{V5}{Baseline covariate}
  \item{V6}{Baseline covariate}
  \item{V7}{Baseline covariate}
}
}
\usage{
data(ex_dta_rct)
}
\description{
Example dataset of a randomized study.
}
\keyword{datasets}
