% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_est.R
\name{plot.PS_RWE_DTA}
\alias{plot.PS_RWE_DTA}
\title{Plot PS distributions}
\usage{
\method{plot}{PS_RWE_DTA}(x, plot_type = c("ps", "balance", "diff"), ...)
}
\arguments{
\item{x}{Class \code{RWE_DWITHPS} created by \code{ps_rwe_*} functions}

\item{plot_type}{Types of plots. \itemize{\item{ps}{PS density plot}
\item{balance}{Covariate balance plot}
\item{diff}{Standardized mean differences, metric = std or astd}}}

\item{...}{Additional parameter for the plot}
}
\description{
S3 method for visualizing PS adjustment
}
