% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{plot.PS_RWE_RST}
\alias{plot.PS_RWE_RST}
\title{Plot estimation results for power prior approach}
\usage{
\method{plot}{PS_RWE_RST}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PS_RWE_RST} that is generated using the
\code{\link{ps_rwe_powerp}}, \code{\link{ps_rwe_compl}}, or
\code{\link{ps_rwe_survkm}} function.}

\item{...}{Additional parameters.}
}
\description{
S3 method plotting estimation results
}
