% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_match.R
\name{plot.PSRWE_DTA_MAT}
\alias{plot.PSRWE_DTA_MAT}
\title{Plot PS distributions}
\usage{
\method{plot}{PSRWE_DTA_MAT}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PSRWE_DTA_MAT} that is generated using
the \code{\link{psrwe_match}} function.}

\item{...}{Parameters for \code{plot.PSRWE_DAT}}
}
\description{
S3 method for visualizing PS adjustment based on matching.
}
\seealso{
\code{\link{plot.PSRWE_DTA}}
}
