% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{plot.PSRWE_RST}
\alias{plot.PSRWE_RST}
\title{Plot estimation results for power prior approach}
\usage{
\method{plot}{PSRWE_RST}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PSRWE_RST} that is generated using the
\code{\link{psrwe_powerp}}, \code{\link{psrwe_compl}}, or
\code{\link{psrwe_survkm}} function.}

\item{...}{Additional parameters.}
}
\description{
S3 method plotting estimation results
}
