% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps.R
\name{plot.RWE_DWITHPS}
\alias{plot.RWE_DWITHPS}
\title{Plot PS distributions}
\usage{
\method{plot}{RWE_DWITHPS}(x, plot_type = c("ps", "balance"), ...)
}
\arguments{
\item{x}{Class \code{RWE_DWITHPS} created by \code{\link{rwe_ps}}}

\item{plot_type}{Types of plots. \itemize{\item{ps}{PS density plot}
\item{balance}{Covariate balance plot}}}

\item{...}{Additional parameter for the plot}
}
\description{
S3 method for visualizing PS adjustment
}
