% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_est.R
\name{plot.RWE_PS_DTA}
\alias{plot.RWE_PS_DTA}
\title{Plot PS distributions}
\usage{
\method{plot}{RWE_PS_DTA}(x, plot_type = c("ps", "balance"), ...)
}
\arguments{
\item{x}{Class \code{RWE_DWITHPS} created by \code{rwe_ps_*} functions}

\item{plot_type}{Types of plots. \itemize{\item{ps}{PS density plot}
\item{balance}{Covariate balance plot}}}

\item{...}{Additional parameter for the plot}
}
\description{
S3 method for visualizing PS adjustment
}
