% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_match.R
\name{plot.RWE_PS_DTA_MAT}
\alias{plot.RWE_PS_DTA_MAT}
\title{Plot PS distributions}
\usage{
\method{plot}{RWE_PS_DTA_MAT}(x, ...)
}
\arguments{
\item{x}{A list of class \code{RWE_PS_DTA_MAT} that is generated using
the \code{\link{rwe_ps_match}} function.}

\item{...}{Parameters for \code{plot.RWE_PS_DAT}}
}
\description{
S3 method for visualizing PS adjustment based on matching.
}
\seealso{
\code{\link{plot.RWE_PS_DTA}}
}
