% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{plot.RWE_PS_RST}
\alias{plot.RWE_PS_RST}
\title{Plot  estimation results for power prior approach}
\usage{
\method{plot}{RWE_PS_RST}(x, ...)
}
\arguments{
\item{...}{Additional parameters.}

\item{object}{A list of class \code{RWE_PS_RST} that is generated using the
\code{\link{rwe_ps_powerp}}, \code{\link{rwe_ps_compl}}, or
\code{\link{rwe_ps_survkm}} function.}
}
\description{

}
