% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survkmplot.R
\name{plot_survkm}
\alias{plot_survkm}
\title{Plot KM at all time points}
\usage{
plot_survkm(
  dta_pskmp,
  conf.int = 0.95,
  conf.type = c("log"),
  xlab = "Time",
  ylab = "Survival Probability",
  ylim = c(0, 1),
  ...
)
}
\arguments{
\item{dta_pskmp}{data}

\item{conf.int}{confidence interval}

\item{conf.type}{confidence type (log as default)}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{ylim}{ylim}

\item{...}{Additional Parameters.}
}
\value{
A KM plot.
}
\description{
Plot KM at all time points
}
