% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_borrow.R
\name{print.PS_RWE_BOR}
\alias{print.PS_RWE_BOR}
\title{Print borrow information}
\usage{
\method{print}{PS_RWE_BOR}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PS_RWE_BOR} that is generated using
the \code{\link{ps_rwe_borrow}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of borrowing
}
\seealso{
\code{\link{ps_rwe_borrow}}
}
