% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_est.R
\name{print.PS_RWE_DTA}
\alias{print.PS_RWE_DTA}
\title{Print PS estimation results}
\usage{
\method{print}{PS_RWE_DTA}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PS_RWE_DTA} that is generated using
the \code{\link{ps_rwe_est}} function.}

\item{...}{Parameters for \code{summery.PS_RWE_DTA}}
}
\description{
Print summary information of PS estimation results
}
\seealso{
\code{\link{summary.PS_RWE_DTA}}
}
