% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_match.R
\name{print.PS_RWE_DTA_MAT}
\alias{print.PS_RWE_DTA_MAT}
\title{Print PS estimation results}
\usage{
\method{print}{PS_RWE_DTA_MAT}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PS_RWE_DTA_MAT} that is generated using
the \code{\link{ps_rwe_match}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of PS estimation results
}
\seealso{
\code{\link{summary.PS_RWE_DTA_MAT}}
}
