% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{print.PS_RWE_RST}
\alias{print.PS_RWE_RST}
\title{Print estimation results}
\usage{
\method{print}{PS_RWE_RST}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PS_RWE_RST} that is generated using the
\code{\link{ps_rwe_powerp}}, \code{\link{ps_rwe_compl}}, or
\code{\link{ps_rwe_survkm}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of outcome mean estimation results
}
\seealso{
\code{\link{summary.PS_RWE_RST}}
}
