% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_outana.R
\name{print.PS_RWE_RST_OUTANA}
\alias{print.PS_RWE_RST_OUTANA}
\title{Print outcome analysis results}
\usage{
\method{print}{PS_RWE_RST_OUTANA}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PS_RWE_RST_OUTANA} that is generated using the
\code{\link{ps_rwe_outana}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of outcome analysis results
}
