% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_borrow.R
\name{print.PSRWE_BOR}
\alias{print.PSRWE_BOR}
\title{Print borrow information}
\usage{
\method{print}{PSRWE_BOR}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PSRWE_BOR} that is generated using
the \code{\link{psrwe_borrow}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of borrowing
}
\seealso{
\code{\link{psrwe_borrow}}
}
