% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_est.R
\name{print.PSRWE_DTA}
\alias{print.PSRWE_DTA}
\title{Print PS estimation results}
\usage{
\method{print}{PSRWE_DTA}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PSRWE_DTA} that is generated using
the \code{\link{psrwe_est}} function.}

\item{...}{Parameters for \code{summery.PSRWE_DTA}}
}
\description{
Print summary information of PS estimation results
}
\seealso{
\code{\link{summary.PSRWE_DTA}}
}
