% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{print.PSRWE_RST}
\alias{print.PSRWE_RST}
\title{Print estimation results}
\usage{
\method{print}{PSRWE_RST}(x, ...)
}
\arguments{
\item{x}{A list of class \code{PSRWE_RST} that is generated using the
\code{\link{psrwe_powerp}}, \code{\link{psrwe_compl}}, or
\code{\link{psrwe_survkm}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of outcome mean estimation results
}
\seealso{
\code{\link{summary.PSRWE_RST}}
}
