% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_borrow.R
\name{print.RWE_PS_BOR}
\alias{print.RWE_PS_BOR}
\title{Print borrow information}
\usage{
\method{print}{RWE_PS_BOR}(x, ...)
}
\arguments{
\item{x}{A list of class \code{RWE_PS_BOR} that is generated using
the \code{\link{rwe_ps_borrow}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of borrowing
}
\seealso{
\code{\link{rwe_ps_borrow}}
}
