% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_est.R
\name{print.RWE_PS_DTA}
\alias{print.RWE_PS_DTA}
\title{Print PS estimation results}
\usage{
\method{print}{RWE_PS_DTA}(x, ...)
}
\arguments{
\item{x}{A list of class \code{RWE_PS_DTA} that is generated using
the \code{\link{rwe_ps_est}} function.}

\item{...}{Parameters for \code{summery.RWE_PS_DTA}}
}
\description{
Print summary information of PS estimation results
}
\seealso{
\code{\link{summary.RWE_PS_DTA}}
}
