% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{print.RWE_PS_RST}
\alias{print.RWE_PS_RST}
\title{Print estimation results}
\usage{
\method{print}{RWE_PS_RST}(x, ...)
}
\arguments{
\item{x}{A list of class \code{RWE_PS_RST} that is generated using the
\code{\link{rwe_ps_powerp}}, \code{\link{rwe_ps_compl}}, or
\code{\link{rwe_ps_survkm}} function.}

\item{...}{Additional parameters}
}
\description{
Print summary information of outcome mean estimation results
}
\seealso{
\code{\link{summary.RWE_PS_RST}}
}
