% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survkmplot.R
\name{print.RWE_PS_RST_PLOT}
\alias{print.RWE_PS_RST_PLOT}
\title{S3 method to print RWE_PS_RST_PLOT}
\usage{
\method{print}{RWE_PS_RST_PLOT}(x, ...)
}
\arguments{
\item{x}{data of RWE_PS_RST_PLOT class}

\item{...}{Additional Parameters.}
}
\description{
S3 method to print RWE_PS_RST_PLOT
}
