% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_complikel.R
\name{ps_rwe_compl}
\alias{ps_rwe_compl}
\title{PS-Integrated Composite Likelihood Estimation}
\usage{
ps_rwe_compl(
  dta_psbor,
  v_outcome = "Y",
  outcome_type = c("continuous", "binary"),
  ...
)
}
\arguments{
\item{dta_psbor}{A class \code{PS_RWE_BOR} object generated by
\code{\link{ps_rwe_borrow}}.}

\item{v_outcome}{Column name corresponding to the outcome.}

\item{outcome_type}{Type of outcomes: \code{continuous} or \code{binary}.}

\item{...}{Parameters for \code{rwe_cl}}
}
\value{
A data frame with class name \code{PS_RWE_RST}. It contains the
    composite estimation of the mean for each stratum as well as the
    jackknife estimation for each subject. The results should be further
    summarized by its S3 method \code{summary}.
}
\description{
Estimate the mean of the outcome based on PS-integrated composite likelihood
approach. Variance is estimated by Jack-Knife method. Applies to the case
when there is only one external data source.
}
\examples{
data(ex_dta)
dta_ps <- ps_rwe_est(ex_dta,
       v_covs = paste("V", 1:7, sep = ""),
       v_grp = "Group",
       cur_grp_level = "current")
ps_borrow <- ps_rwe_borrow(total_borrow = 30, dta_ps)
rst       <- ps_rwe_compl(ps_borrow, v_outcome = "Y_Con")
rst

}
