% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_outana.R
\name{ps_rwe_outana}
\alias{ps_rwe_outana}
\title{Outcome Analysis for PS-Integrated Estimation

Report outcome analysis for the PS-integrated approach.}
\usage{
ps_rwe_outana(dta_psrst)
}
\arguments{
\item{dta_psrst}{a returned object with class \code{PS_RWE_EST}}

\item{...}{other options}
}
\value{
A list with class name \code{PS_RWE_EST_OUTANA}.
}
\description{
Outcome Analysis for PS-Integrated Estimation

Report outcome analysis for the PS-integrated approach.
}
\details{
This function is mainly for summarizing and reporting the
    outcome analysis for the PS-integrated estimation.
    The input \code{dta_psrst} can be generated from the functions
    \code{\link{ps_rwe_powerp}}, \code{\link{ps_rwe_compl}},
    \code{\link{ps_rwe_survkm}}, \code{\link{ps_rwe_ci}}, and.
    \code{\link{ps_rwe_infer}}.
}
\examples{
data(ex_dta)
dta_ps <- ps_rwe_est(ex_dta,
       v_covs = paste("V", 1:7, sep = ""),
       v_grp = "Group",
       cur_grp_level = "current")
ps_borrow <- ps_rwe_borrow(total_borrow = 30, dta_ps)
ps_rst <- ps_rwe_compl(ps_borrow, v_outcome = "Y_Con")
rst <- ps_rwe_outana(ps_rst)
rst

}
