% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{psrwe_powerp}
\alias{psrwe_powerp}
\title{Get posterior samples based on PS-power prior approach}
\usage{
psrwe_powerp(
  dta_psbor,
  v_outcome = "Y",
  outcome_type = c("continuous", "binary"),
  prior_type = c("fixed", "random"),
  prioronly = FALSE,
  ...,
  seed = NULL
)
}
\arguments{
\item{dta_psbor}{A class \code{PSRWE_BOR} object generated by
\code{\link{psrwe_borrow}}.}

\item{v_outcome}{Column name corresponding to the outcome.}

\item{outcome_type}{Type of outcomes: \code{continuous} or \code{binary}.}

\item{prior_type}{Whether treat power parameter as fixed (\code{fixed}) or
fully Bayesian (\code{random}).}

\item{prioronly}{Whether only obtain power prior by excluding the current
study data).}

\item{...}{extra parameters for calling function \code{\link{rwe_stan}}.}

\item{seed}{Random seed.}
}
\value{
A class \code{PSRWE_RST} list with the following objects

\describe{

    \item{Observed}{Observed mean and SD of the outcome by group, arm
    and stratum}

    \item{Control}{A list of estimated mean and SD of the outcome by stratum
    in the control arm}

    \item{Treatment}{A list of estimated mean and SD of the outcome by
    stratum in the treatment arm for RCT}

    \item{Effect}{A list of estimated mean and SD of the treatment effect by
    stratum for RCT}

    \item{Borrow}{Borrowing information from \code{dta_psbor}}

    \item{stan_rst}{Result from STAN sampling}
}
}
\description{
Draw posterior samples of the parameters of interest for the PS-power prior
approach
}
\examples{

\donttest{
data(ex_dta)
dta_ps <- psrwe_est(ex_dta,
       v_covs = paste("V", 1:7, sep = ""),
       v_grp = "Group",
       cur_grp_level = "current")
ps_borrow <- psrwe_borrow(total_borrow = 30, dta_ps)
rst <- psrwe_powerp(ps_borrow, v_outcome = "Y_Con", seed = 123)}

}
