% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survcoxphwa.R
\name{rwe_coxphwa}
\alias{rwe_coxphwa}
\title{The coxph estimation (weighted average approach)}
\usage{
rwe_coxphwa(
  dta_cur,
  dta_ext,
  dta_cur_trt,
  n_borrow = 0,
  stderr_method = "naive"
)
}
\arguments{
\item{dta_cur}{Matrix of time and event from a PS stratum in current study
(control arm only)}

\item{dta_ext}{Matrix of time and event from a PS stratum in external data
source (control arm only)}

\item{dta_cur_trt}{Matrix of time and event from a PS stratum in current
study (treatment arm only)}

\item{n_borrow}{Number of subjects to be borrowed}

\item{stderr_method}{Method for computing StdErr (available for naive only)}
}
\value{
Estimation of coxph estimates
}
\description{
Estimate coxph estimates for a single PS stratum
}
