% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_confintv.R
\name{rwe_ps_ci}
\alias{rwe_ps_ci}
\title{Confidence Interval for PS-Integrated Estimation

Estimate the confidence interval for he PS-integrated approach.}
\usage{
rwe_ps_ci(
  dta_psrst,
  method_ci = c("wald", "wilson"),
  conf_type = c("log_log", "plain"),
  conf_int = 0.95,
  ...
)
}
\arguments{
\item{dta_psrst}{a returned object with class \code{RWE_PS_EST}}

\item{method_ci}{a method name for confidence interval (default Wald)}

\item{conf_type}{a type name of transformation for the confidence interal
of PSKM approach (default log_log)}

\item{conf_int}{a two-sided level of confidence limits (default 0.95)}

\item{...}{other options}
}
\value{
A list with class name \code{RWE_PS_EST}.
}
\description{
Confidence Interval for PS-Integrated Estimation

Estimate the confidence interval for he PS-integrated approach.
}
\examples{
\donttest{

data(ex_dta)
dta_ps <- rwe_ps_est(ex_dta,
       v_covs = paste("V", 1:7, sep = ""),
       v_grp = "Group",
       cur_grp_level = "current")
ps_borrow <- rwe_ps_borrow(total_borrow = 30, dta_ps)
ps_rst <- rwe_ps_compl(ps_borrow, v_outcome = "Y_Con")
rst <- rwe_ps_ci(ps_rst)
}

}
