% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_ps_est.R
\name{rwe_ps_est}
\alias{rwe_ps_est}
\title{Estimate propensity scores}
\usage{
rwe_ps_est(
  data,
  ps_fml = NULL,
  ps_method = c("logistic", "randomforest"),
  v_covs = "V1",
  v_grp = "Group",
  cur_grp_level = 1,
  v_arm = NULL,
  ctl_arm_level = NULL,
  nstrata = 5,
  ...
)
}
\arguments{
\item{data}{Data frame with group assignment and covariates.}

\item{ps_fml}{Propensity score (PS) formula. If \code{NULL}, all covariates
will be included in the PS model in a linear form.}

\item{ps_method}{Method to calculate propensity scores. Can be set to
\code{logistic} for logistic regression or \code{randomforest} for a
random forest approach.}

\item{v_grp}{Column name corresponding to group assignment.}

\item{cur_grp_level}{Group level for the current study. Default is
\code{cur_grp_level = 1}. Ignored for single arm studies.}

\item{ctl_arm_level}{Arm level for the control arm. Ignored for single-arm
studies.}

\item{nstrata}{Number of PS strata to be created.}

\item{...}{Additional parameters for calculating the propensity score to be
used in \code{randomForest} or \code{glm} .}
}
\value{
A list of class \code{RWE_PS_DAT} with items:

\itemize{
  \item{data}{Original data with column \code{_ps_} for estimated PS scores
              and \code{_strata_} for PS stratum added.}
  \item{ps_fml}{PS formula for estimated PS scores.}
  \item{is_rct}{Whether the current study is a randomized study.}
  \item{nstrata}{Number of strata.}
}
}
\description{
Estimate propensity scores using logistic regression or random
  forest model.
}
\examples{
data(ex_dta)
rwe_ps_est(ex_dta,
       v_covs = paste("V", 1:7, sep = ""),
       v_grp = "Group",
       cur_grp_level = "current")

}
