% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survkmplot.R
\name{rwe_ps_survkmci}
\alias{rwe_ps_survkmci}
\title{Obtain CI for survkm plot.}
\usage{
rwe_ps_survkmci(dta_pskmp, conf_int = 0.95, conf_type = c("log-log", "plain"))
}
\arguments{
\item{dta_pskmp}{data of RWE_PS_RST_PLOT class}

\item{conf_int}{level of confidence interval}

\item{conf_type}{type of confidence interval (log-log as default)}
}
\value{
A data frame with class name \code{RWE_PS_RST_PLOT}. It contains the
    composite estimation of the mean for overall and each stratum.
    It also contains confidence interval for overall mean.
}
\description{
Obtain CI for survkm plot.
}
