% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survkmplot.R
\name{rwe_ps_survkmplot}
\alias{rwe_ps_survkmplot}
\title{PS-Integrated Kaplan-Meier Estimation Plot}
\usage{
rwe_ps_survkmplot(dta_psbor, v_time = "time", v_event = "event", ...)
}
\arguments{
\item{dta_psbor}{A class \code{RWE_PS_BOR} object generated by
\code{\link{rwe_ps_borrow}}.}

\item{v_time}{Column name corresponding to event time}

\item{v_event}{Column name corresponding to event status}

\item{...}{Additional Parameters.}
}
\value{
A data frame with class name \code{RWE_PS_RST}. It contains the
    composite estimation of the mean for overall and each stratum.
}
\description{
Plot the mean of a survival outcome at all distinctive time points based on
PS-integrated Kaplan-Meier approach. Variance is estimated by usual Greenwood
method. Applies to the case when there is only one external data source.
}
