% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_tools.R
\name{rweCut}
\alias{rweCut}
\title{Cut a sequence of numbers into bins with equal numbers in each bin}
\usage{
rweCut(x, y = x, breaks = 5)
}
\description{
Cut a sequence of numbers into bins with equal numbers in each bin
}
