% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_tools.R
\name{rweDist}
\alias{rweDist}
\title{Compute distance from F0 to F1}
\usage{
rweDist(sample.F0, sample.F1, n.bins = 10, type = c("kl", "ovl"),
  epsilon = 10^-6)
}
\arguments{
\item{n.bins}{number of bins for KL computation}

\item{type}{type of distances. ovl: overlapping coefficient, kl:
1/(1+Kullback-Leibler divergence)}

\item{epsilon}{small integer for Dirichlet smoothing}
}
\value{
a vector with the number of samples in group 0, the number of samples
    in group 1, and 1/(1+KL divergence) from group 0 to group 1 when type is
    kl, or the overlapping coefficient when type is ovl
}
\description{
Compute distance from F0 to F1
}
