% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_tools.R
\name{rweFreqTbl}
\alias{rweFreqTbl}
\title{Generate frequency table for factor columns}
\usage{
rweFreqTbl(data, var.groupby, vars = NULL)
}
\value{
a vector with the number of samples in group 0, the number of samples
    in group 1, and the KL divergence from group 0 to group 1
}
\description{
Generate frequency table for factor columns
}
