% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweGetBinInt}
\alias{rweGetBinInt}
\title{Get intercept for a binary outcome.}
\usage{
rweGetBinInt(..., regCoeff, nPat = 5e+05, xbeta = NULL, bin.mu = 0.5)
}
\arguments{
\item{...}{Parameters for simulating covariates by function
\code{\link{rweXBeta}} and \code{\link{rweSimuCov}}}

\item{regCoeff}{regression coefficients}

\item{nPat}{number of patients}

\item{bin.mu}{mean of the binary outcomes used to compute b0}
}
\value{
standard error or the random error term
}
\description{
The binary outcome may be an outcome or a treatment assignment.
}
