% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweGetYSig}
\alias{rweGetYSig}
\title{Compute standard error of the random error}
\usage{
rweGetYSig(..., nPat = 5e+05, xbeta = NULL, sig2Ratio = 1)
}
\arguments{
\item{nPat}{number of patients}
}
\value{
mean of xbeta and standard error or the random error term
}
\description{
Compute standard error of the random error
}
