% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_present.R
\name{rwePlotUnbalance}
\alias{rwePlotUnbalance}
\title{Plot unbalance in covariates}
\usage{
rwePlotUnbalance(data.unb, var.x = "Diff", var.group = NULL,
  xlim = NULL, ylim = NULL, title = "",
  f.grid = formula("V~Study"), adjust = 1)
}
\description{
Plot unbalance in covariates
}
