% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_stan.R
\name{rwePowerDrawPost}
\alias{rwePowerDrawPost}
\title{Get Posterior for all stratum}
\usage{
rwePowerDrawPost(data.ps, v.outcome = "Y", As = 0,
  type = c("continuous", "binary"), ...)
}
\arguments{
\item{data.ps}{class DWITHPS data frame}

\item{...}{extra parameters for calling function \code{\link{rweSTAN}}}

\item{A}{power of the power prior for each strata}
}
\description{
Get Posterior for all stratum
}
