% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_ps.R
\name{rwePS}
\alias{rwePS}
\title{Get propensity scores}
\usage{
rwePS(data, ps.fml = NULL, v.grp = "group", v.covs = "V1",
  d1.grp = 1, nstrata = 5, ..., caliper = 0)
}
\arguments{
\item{...}{parameters to get propensity scores}
}
\description{
Get propensity scores
}
