% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuCombine}
\alias{rweSimuCombine}
\title{Combining simulation results}
\usage{
rweSimuCombine(lst.rst, fun = mean, ignore.error = TRUE, ...)
}
\arguments{
\item{lst.rst}{List of simulation results. Each element represents a
replication}
}
\description{
Combining simulation results
}
