% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuError}
\alias{rweSimuError}
\title{Simulate random errors}
\usage{
rweSimuError(nPat, error.type = c("normal", "skewed"), ysig = 1,
  skew.n = NULL, skew.p = NULL, skew.noise = 1e-04, ...)
}
\arguments{
\item{nPat}{number of patients}

\item{ysig}{standard error of the random error}

\item{skew.n}{parameter of negative bionomial distribution}

\item{skew.p}{parameter of negative bionomial distribution}
}
\description{
Simulate random errors
}
