% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuSingleArm}
\alias{rweSimuSingleArm}
\title{Simulate outcomes for a single arm study}
\usage{
rweSimuSingleArm(nPat, muCov, sdCov, corCov, regCoeff, mix.phi = 1,
  cov.breaks = NULL, fmla = NULL, type = c("continuous", "binary"),
  ysig = NULL, sig2Ratio = 1, b0 = NULL, bin.mu = 0.5, ...)
}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}

\item{regCoeff}{regression coefficients}

\item{mix.phi}{weight in mixture model}

\item{cov.breaks}{breaks to cut covaraites into categories. If numeric, the
same cuts will be applied to all covariates. If list, each covariates
will be categorized based on its own breaks. If NULL, keep continuous.}

\item{type}{distributions of the random error}

\item{ysig}{standard error of the random error}

\item{b0}{intercept in regession model}

\item{bin.mu}{mean of the binary outcomes used to compute b0}
}
\description{
Simulate outcomes for a single arm study
}
