% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweSimuTwoArm}
\alias{rweSimuTwoArm}
\title{Simulate continuous outcomes for a two arm study}
\usage{
rweSimuTwoArm(nPat, muCov, sdCov, corCov, trt.effect = 0, regCoeff.y,
  regCoeff.z = 0, mix.phi = 1, cov.breaks = NULL, fmla.y = NULL,
  fmla.z = NULL, ysig = NULL, b0 = NULL, z1.p = 0.5,
  sig2Ratio = 2, ..., do.simu = TRUE)
}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}

\item{mix.phi}{weight in mixture model}

\item{cov.breaks}{breaks to cut covaraites into categories. If numeric, the
same cuts will be applied to all covariates. If list, each covariates
will be categorized based on its own breaks. If NULL, keep continuous.}

\item{ysig}{standard error of the random error}

\item{b0}{intercept in regession model}
}
\description{
Simulate continuous outcomes for a two arm study
}
