% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_stan.R
\name{rweSummaryPost}
\alias{rweSummaryPost}
\title{Summary Posterior theta}
\usage{
rweSummaryPost(post.theta, true.theta, quants = c(0.025, 0.975),
  weights = 1)
}
\arguments{
\item{post.theta}{posterior samples from STAN}

\item{true.theta}{true value of theta}

\item{quants}{quantiles}

\item{weights}{number of subjects in each strata as the weight}
}
\description{
Summary Posterior theta
}
