% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_tools.R
\name{rweUnbalance}
\alias{rweUnbalance}
\title{Get unbalance in baseline covariates}
\usage{
rweUnbalance(nPat, ..., pts = NULL, covs = NULL, diff = TRUE,
  var.group = "Z", cov.pattern = "^V[0-9]+$")
}
\arguments{
\item{nPat}{number of patients}

\item{diff}{If TRUE, get the difference in covariates between groups.
Otherwise, get covariates for each group separately}

\item{var.group}{Column name in pts that corresponds to treat group}
}
\value{
If diff is TRUE, return a dataset with columns V and Diff. Otherwise,
    return a dataset with columns V, Z and Value. In both cases, column V
    represents the name of covariates.
}
\description{
Get unbalance in baseline covariates
}
