% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_simu.R
\name{rweXBeta}
\alias{rweXBeta}
\title{Simulate X multiplied by Beta}
\usage{
rweXBeta(..., regCoeff, cov.x = NULL, fmla = NULL)
}
\arguments{
\item{...}{Parameters for simulating covariates by function
\code{\link{rweSimuCov}}}

\item{regCoeff}{regression coefficients}
}
\description{
Simulate X multiplied by Beta
}
