% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwe_package.R
\name{simupara}
\alias{simupara}
\title{Parameters for simulating data}
\arguments{
\item{nPat}{number of patients}

\item{muCov}{mean vector of covariates}

\item{sdCov}{standard deviation vector of covariates}

\item{corCov}{correlation of covariates}

\item{regCoeff}{regression coefficients}

\item{mix.phi}{weight in mixture model}

\item{cov.breaks}{breaks to cut covaraites into categories. If numeric, the
same cuts will be applied to all covariates. If list, each covariates
will be categorized based on its own breaks. If NULL, keep continuous.}

\item{type}{distributions of the random error}

\item{ysig}{standard error of the random error}

\item{skew.n}{parameter of negative bionomial distribution}

\item{skew.p}{parameter of negative bionomial distribution}

\item{b0}{intercept in regession model}

\item{bin.mu}{mean of the binary outcomes used to compute b0}

\item{formula.z}{formula of the treatement assignment model. No intercept
term.}

\item{formula.y}{formula of the outcome model. No intercept term.}

\item{trial.data}{existing clinical trial data}

\item{group}{column referring to arm in the existing dataset}

\item{outcome}{column referring to outcome in the existing dataset}

\item{trt.effec}{true treatment effect to be set}

\item{with.replacement}{sample with or without replacement from the existing
dataset}

\item{keep.group}{sample ignore arm assignment in the existing dataset}
}
\description{
Parameters for simulating data
}
