% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{summary.PS_RWE_RST}
\alias{summary.PS_RWE_RST}
\title{Summarize overall estimation results}
\usage{
\method{summary}{PS_RWE_RST}(object, ...)
}
\arguments{
\item{object}{A list of class \code{PS_RWE_RST} that is generated using the
\code{\link{ps_rwe_powerp}}, \code{\link{ps_rwe_compl}}, or
\code{\link{ps_rwe_survkm}} function.}

\item{...}{Additional parameters.}
}
\value{
A list with data frames for the borrowing and estimation results.
}
\description{
S3 method summarizing overall estimation results
}
