% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{summary.PSRWE_RST}
\alias{summary.PSRWE_RST}
\title{Summarize overall estimation results}
\usage{
\method{summary}{PSRWE_RST}(object, ...)
}
\arguments{
\item{object}{A list of class \code{PSRWE_RST} that is generated using the
\code{\link{psrwe_powerp}}, \code{\link{psrwe_compl}}, or
\code{\link{psrwe_survkm}} function.}

\item{...}{Additional parameters.}
}
\value{
A list with data frames for the borrowing and estimation results.
}
\description{
S3 method summarizing overall estimation results
}
