% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_outana.R
\name{summary.PSRWE_RST_OUTANA}
\alias{summary.PSRWE_RST_OUTANA}
\title{Summary outcome analysis results}
\usage{
\method{summary}{PSRWE_RST_OUTANA}(object, pred_tps = NULL, ...)
}
\arguments{
\item{object}{A list of class \code{PSRWE_RST_OUTANA} that is generated using the
\code{\link{psrwe_outana}} function.}

\item{pred_tps}{Specified time points}

\item{...}{Additional parameters}
}
\description{
Summary information of outcome analysis results
}
