% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_complikel.R
\name{summary.RWE_CLRST}
\alias{summary.RWE_CLRST}
\title{Summary of Composite Likelihood Estimation}
\usage{
\method{summary}{RWE_CLRST}(object, ...)
}
\arguments{
\item{object}{A class \code{RWE_CLRST} object}

\item{...}{Extra arguments}
}
\value{
Summary of composite likelihood estimation
}
\description{
Summarize the result in a class \code{RWE_CLRST} objects
generated by \code{\link{rwe_ps_cl}}.
}
