% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{summary.RWE_POWERPRST}
\alias{summary.RWE_POWERPRST}
\title{Summary of PS-power prior Estimation}
\usage{
\method{summary}{RWE_POWERPRST}(object, ...)
}
\arguments{
\item{object}{A class \code{RWE_POWERPRST} object}

\item{...}{Extra arguments}
}
\value{
Summary of composite likelihood estimation
}
\description{
Summarize the result in a class \code{RWE_POWERPRST} objects
generated by \code{\link{rwe_ps_cl}}.
}
