% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_powerprior.R
\name{summary.RWE_PS_RST}
\alias{summary.RWE_PS_RST}
\title{Summarize overall estimation results}
\usage{
\method{summary}{RWE_PS_RST}(object, ...)
}
\arguments{
\item{object}{A list of class \code{RWE_PS_RST} that is generated using the
\code{\link{rwe_ps_powerp}}, \code{\link{rwe_ps_compl}}, or
\code{\link{rwe_ps_survkm}} function.}

\item{...}{Additional parameters.}
}
\value{
A list with data frames for the borrowing and estimation results.
}
\description{
S3 method summarizing overall estimation results
}
