% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{tkCallFun}
\alias{tkCallFun}
\title{Call function by its name organized as a vector}
\usage{
tkCallFun(vec, ...)
}
\arguments{
\item{vec}{function names as a vector}

\item{...}{Parameters needed for the actual function}
}
\description{
Call function by its name organized as a vector
}
