% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{tkMakeLocal}
\alias{tkMakeLocal}
\title{Import objects in a list into a designated environment}
\usage{
tkMakeLocal(alist, dest.env)
}
\arguments{
\item{alist}{list of objects}

\item{dest.env}{designated environment}
}
\description{
Import objects in a list into a designated environment
}
