% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_package.R
\name{ex_dta}
\alias{ex_dta}
\title{Example dataset}
\format{
A dataframe with the following variables:
\itemize{
  \item{Group}{current, rwd}
  \item{Y}{Binary outcome}
  \item{V1-V7}{Covariates}
}
}
\description{
Example dataset of a single arm study
}
