% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_package.R
\name{ex_dta_rct}
\alias{ex_dta_rct}
\title{Example dataset}
\format{
A dataframe with the following variables:
\itemize{
  \item{Group}{0,1}
  \item{Arm}{0, 1}
  \item{Y}{Continuous outcome}
  \item{V1-V7}{Covariates}
}
}
\description{
Example dataset of a randomized study
}
