% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_package.R
\docType{data}
\name{ex_dta_rct}
\alias{ex_dta_rct}
\title{Example dataset

Example dataset of a randomized study.}
\format{
A data frame with the following variables:
\itemize{
  \item{Group}{current, rwd}
  \item{Arm}{control, treatment}
  \item{Y_Con}{Continuous outcome}
  \item{V1-V7}{Covariates}
}
}
\usage{
data(ex_dta_rct)
}
\description{
Example dataset

Example dataset of a randomized study.
}
\keyword{datasets}
