% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survkmplot.R
\name{plot_survkm}
\alias{plot_survkm}
\title{Plot KM at all time points}
\usage{
plot_survkm(dta_pskmp, xlab = "Time", ylab = "Survival Probability", ...)
}
\arguments{
\item{dta_pskmp}{data of RWE_PS_RST_PLOT class}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{...}{Additional Parameters.}

\item{ylim}{ylim}
}
\value{
A KM plot.
}
\description{
Plot KM at all time points
}
