% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_outana.R
\name{print.PSRWE_RST_OUTANA}
\alias{print.PSRWE_RST_OUTANA}
\title{Print outcome analysis results}
\usage{
\method{print}{PSRWE_RST_OUTANA}(x, show_details = FALSE, show_rct = FALSE, ...)
}
\arguments{
\item{x}{A list of class \code{PSRWE_RST_OUTANA} that is generated using the
\code{\link{psrwe_outana}} function.}

\item{show_details}{print out more observed summary}

\item{show_rct}{print out more analysis summary for RCT arms}

\item{...}{Additional parameters}
}
\description{
Print summary information of outcome analysis results
}
