% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survcoxphsp.R
\name{psrwe_survcoxphsp}
\alias{psrwe_survcoxphsp}
\title{PS-integrated Cox proportional hazard method for
comparing time-to-event outcomes (same proportion approach)}
\usage{
psrwe_survcoxphsp(
  dta_psbor,
  v_time = "time",
  v_event = "event",
  stderr_method = c("naive", "jk", "sjk", "cjk", "sbs", "cbs"),
  ...
)
}
\arguments{
\item{dta_psbor}{A class \code{PSRWE_BOR} object generated by
\code{\link{psrwe_borrow}}.}

\item{v_time}{Column name corresponding to event time}

\item{v_event}{Column name corresponding to event status}

\item{stderr_method}{Method for computing StdErr (see Details)}

\item{...}{Additional Parameters}
}
\value{
A data frame with class name \code{PSRWE_RST_TESTANA}.
    It contains the test statistics as well as the
    Jackknife estimation. The results can be further
    summarized by its S3 method \code{summary}.
    The results can be also analyzed by \code{psrwe_outana} for outcome
    analysis and inference.
}
\description{
Cox proportional hazard (coxph) method evaluates two-arm RCT via
PS-integrated method (same proportion approach) via
stratified proportional hazard model.
Apply to the case when there is only one external data source and
two-arm RCT.
}
\details{
\code{stderr_method} includes \code{naive} as default which
    mostly follows the calculation provided by
    \code{survival::coxph(..., robust = TRUE)}, and
    \code{jk}, \code{sjk}, \code{cjk}, \code{sbs}, \code{cbs}, and
    \code{none}. See \code{\link{stderr_method}} for details.

    Naive approach calculates log hazard ratio by each stratum, then
    takes weighted average of all stratum-specific estimates as other
    PS-integrated methods in \pkg{psrwe}.

    The PS-integrated coxph method optimizes the composite partial
    likelihood to obtain the point estimate of hazard ratio between
    two arms.
    The implementation of the optimization is equivalent to estimate the
    hazard ratio of the weighted and stratified coxph model
    via \code{survival::coxph()}.
}
\examples{
data(ex_dta_rct)
dta_ps_rct <- psrwe_est(ex_dta_rct,
                        v_covs = paste("V", 1:7, sep = ""),
                        v_grp = "Group", cur_grp_level = "current",
                        v_arm = "Arm", ctl_arm_level = "control",
                        ps_method = "logistic", nstrata = 5,
                        stra_ctl_only = FALSE)
ps_bor_rct <- psrwe_borrow(dta_ps_rct, total_borrow = 30)
rst_coxphsp <- psrwe_survcoxphsp(ps_bor_rct,
                                 v_time = "Y_Surv",
                                 v_event = "Status")
rst_coxphsp

}
