% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survkm.R
\name{psrwe_survkm}
\alias{psrwe_survkm}
\title{PS-integrated Kaplan-Meier estimation}
\usage{
psrwe_survkm(
  dta_psbor,
  pred_tp,
  v_time = "time",
  v_event = "event",
  stderr_method = c("naive", "jk", "sjk", "cjk", "sbs", "cbs", "none"),
  ...
)
}
\arguments{
\item{dta_psbor}{A class \code{PSRWE_BOR} object generated by
\code{\link{psrwe_borrow}}.}

\item{pred_tp}{A numeric value corresponding to time of interest
(e.g., 365 days or 1 year)}

\item{v_time}{Column name corresponding to event time}

\item{v_event}{Column name corresponding to event status}

\item{stderr_method}{Method for computing StdErr, see Details}

\item{...}{Additional Parameters}
}
\value{
A data frame with class name \code{PSRWE_RST}. It contains the
    composite estimation of the mean for each stratum as well as the
    Jackknife estimation. The results can be further
    summarized by its S3 method \code{summary}.
    The results can be also analyzed by \code{psrwe_outana} for outcome
    analysis and inference.
}
\description{
Estimate the mean of a survival outcome at a given time point based on
PS-integrated Kaplan-Meier approach. Variance can be estimated by Jackknife
methods. Apply to the case when there is only one external data source.
}
\details{
\code{stderr_method} includes \code{naive} as default which
    mostly follows Greenwood formula,
    \code{jk} using Jackknife method within each stratum,
    \code{sjk} using simple Jackknife method for combined estimates
    such as point estimates in single arm or treatment effects in RCT, or
    \code{cjk} for complex Jackknife method including refitting PS model,
    matching, trimming, calculating borrowing parameters, and
    combining overall estimates.
    Note that \code{sjk} may take a while longer to finish and
    \code{cjk} will take even much longer to finish.
    The \code{sbs} and \code{cbs} are for simple and complex Bootstrap
    methods (\code{n_bootstrap = 200} as default).
}
\examples{
data(ex_dta)
dta_ps <- psrwe_est(ex_dta,
       v_covs = paste("V", 1:7, sep = ""),
       v_grp = "Group",
       cur_grp_level = "current")
ps_borrow <- psrwe_borrow(total_borrow = 30, dta_ps)
rst       <- psrwe_survkm(ps_borrow,
                          pred_tp = 365,
                          v_time = "Y_Surv",
                          v_event = "Status")
rst

}
