% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survlrk.R
\name{psrwe_survlrk}
\alias{psrwe_survlrk}
\title{PS-Integrated Log-Rank Test For Comparing Time-to-event Outcomes}
\usage{
psrwe_survlrk(
  dta_psbor,
  v_time = "time",
  v_event = "event",
  pred_tp = 1,
  stderr_method = c("naive", "jk", "jkoverall"),
  ...
)
}
\arguments{
\item{dta_psbor}{A class \code{PSRWE_BOR} object generated by
\code{\link{psrwe_borrow}}.}

\item{v_time}{Column name corresponding to event time}

\item{v_event}{Column name corresponding to event status}

\item{pred_tp}{Time of interest (e.g., 1 year)}

\item{stderr_method}{Method for computing StdErr (see Details)}

\item{...}{Additional Parameters}
}
\value{
A data frame with class name \code{PSRWE_RST_TESTANA}.
    It contains the test statistics of each stratum as well as the
    jackknife estimation. The results can be further
    summarized by its S3 method \code{summary}.
    The results can be also analyzed by \code{psrwe_outana} for outcome
    analysis and inference.
}
\description{
Log-rank test evaluates two-arm RCT for up to a given time point.
Variance can be estimated by Jackknife methods.
Applies to the case when there is only one external data source and
two-arm RCT.
}
\details{
\code{stderr_method} include \code{naive} as default which
    mostly follows Greenwood formula (plug-in),
    \code{jk} using Jackknife method within each stratum, or
    \code{jkoverall} using Jackknife method for overall/combined estimates.
    Note that \code{jkoverall} may take a while longer to finish.
}
\examples{
data(ex_dta_rct)
dta_ps_rct <- psrwe_est(ex_dta,
                        v_covs = paste("V", 1:7, sep = ""),
                        v_grp = "Group", cur_grp_level = "current",
                        v_arm = "Arm", ctl_arm_level = "control",
                        ps_method = "logistic", nstrata = 5,
                        stra_ctl_only = FALSE)
ps_bor_rct <- psrwe_borrow(dta_ps_rct, total_borrow = 30)
rst_lrk <- psrwe_survlrk(ps_bor_rct,
                         v_time = "Y_Surv",
                         v_event = "Status")
rst_lrk

}
