% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survcoxphsp.R
\name{rwe_coxphsp}
\alias{rwe_coxphsp}
\title{The coxph estimation (same proportion approach)}
\usage{
rwe_coxphsp(
  dta_cur,
  dta_ext,
  dta_cur_trt,
  n_borrow = 0,
  stderr_method = "naive"
)
}
\arguments{
\item{dta_cur}{Matrix of time and event from a PS stratum in current study
(control arm only)}

\item{dta_ext}{Matrix of time and event from a PS stratum in external data
source (control arm only)}

\item{dta_cur_trt}{Matrix of time and event from a PS stratum in current
study (treatment arm only)}

\item{n_borrow}{Number of subjects to be borrowed}

\item{stderr_method}{Method for computing StdErr (available for naive only)}
}
\value{
Estimation of overall coxph estimate
}
\description{
Estimate overall coxph estimate with all strata together
}
