% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psrwe_survkmplot.R
\name{summary.RWE_PS_RST_PLOT}
\alias{summary.RWE_PS_RST_PLOT}
\title{S3 method to summary RWE_PS_RST_PLOT}
\usage{
\method{summary}{RWE_PS_RST_PLOT}(object, ...)
}
\description{
S3 method to summary RWE_PS_RST_PLOT
}
